% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooleDS.o.R
\name{BooleDS.o}
\alias{BooleDS.o}
\title{BooleDS.o}
\usage{
BooleDS.o(V1.name = NULL, V2.name = NULL, Boolean.operator.n = NULL,
  na.assign.text, numeric.output = TRUE)
}
\arguments{
\item{V1.name}{A character string specifying the name of the vector to which the Boolean operator is to be applied}

\item{V2.name}{A character string specifying the name of the vector or scalar to which <V1> is to be compared.}

\item{numeric.output}{a TRUE/FALSE indicator defaulting to TRUE determining whether the final output variable
should be of class numeric (1/0) or class logical (TRUE/FALSE).}

\item{Boolean.operator}{A character string specifying one of six possible Boolean operators: '==', '!=', '>', '>=',
'<', '<='}

\item{na.assign}{A character string taking values 'NA', '1' or '0'. If 'NA' then any NA values in the
input vector remain as NAs in the output vector. If '1' or '0' NA values in the input vector are
all converted to 1 or 0 respectively.#' @return the levels of the input variable.}
}
\description{
Converts the individual elements of a vector or other object into Boolean indicators.
}
\details{
The function converts the input vector into Boolean indicators.
}
\author{
Burton PR
}
