% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlotDS.o.R
\name{BoxPlotDS.o}
\alias{BoxPlotDS.o}
\title{Computes the statistics of the input variable required for creating a boxplot.}
\usage{
BoxPlotDS.o(x)
}
\arguments{
\item{x}{the input (numeric) vector for which the boxplot is desired.}
}
\value{
a list which includes the objects stats, n, conf, out and coef
}
\description{
This function produces all the information required to plot a boxplot.
}
\details{
The output information is a list with 5 objects named stats, n, conf,
out and coef. Object stats is a 5-length vector which contains the extreme of 
the lower whisker (equal to the 5% quantile), the lower hinge (equal to the 25%
quantile), the median, the upper hinge (equal to the 75% quantile) and the
extreme of the upper whisker (equal to the 95% quantile). Object n is the number
of all non-NA observations in the dataset. Object conf gives the lower and upper
extremes of the notch. Object out gives the number of outliers which are below the
lower whisker and the number of outliers which are above the upper whisker. Object
coef gives the range of whiskers (how far they extend out from the box).
The range of whiskers is set equal to (95% quantile-75%quantile)/(interquartile range).
This confirms that the extreme of the lower whisker is equal to the 5% quantile
and the extreme of the upper whisker is equal to the 95% quantile (avoiding any 
possible disclosure).
}
\author{
Avraam D
}
