% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids1dDS.o.R
\name{centroids1dDS.o}
\alias{centroids1dDS.o}
\title{Calculates the centroid of each n nearest neighbours}
\usage{
centroids1dDS.o(x, n)
}
\arguments{
\item{x}{the name of a numeric vector, the x-variable.}

\item{n}{the number of the nearest neghbours for which their centroid is calculated.}
}
\value{
a vector with the centroids in 1D
}
\description{
This function calculates the centroids for each n nearest neighbours.
}
\details{
The function finds the n-1 nearest neighbours of each data point in the 1-dimensional space.
The nearest neighbours are the data points with the minimum distances (in 1D, the distance is given 
by the difference between two values) from the point of interest. Each point of interest and its n-1
nearest neighbours are then used for the calculation of the centroid of those n points. The centroid
is the average value of the n nearest neighbours. The centroids are returned to the client side function
and can be used for the generation of non-disclosive graphs (e.g. histograms, boxplots, etc).
}
\author{
Avraam D
}
