% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids2dDS.o.R
\name{centroids2dDS.o}
\alias{centroids2dDS.o}
\title{Calculates the coordinates of the centroid of each n nearest neighbours}
\usage{
centroids2dDS.o(x, y, n)
}
\arguments{
\item{x}{the name of a numeric vector, the x-variable.}

\item{y}{the name of a numeric vector, the y-variable.}

\item{n}{the number of the nearest neghbours for which their centroid is calculated.}
}
\value{
a list with the x and y coordinates of the centroids
}
\description{
This function calculates the coordinates of the centroids for each n nearest neighbours.
}
\details{
The function finds the n-1 nearest neighbours of each data point in a 2-dimensional space.
The nearest neighbours are the data points with the minimum Euclidean distances from the point of 
interest. Each point of interest and its n-1 nearest neighbours are then used for the calculation
of the coordinates of the centroid of those n points. Centroid here is referred to the centre of mass,
i.e. the x-coordinate of the centroid is the average value of the x-coordinates of the n nearest
neighbours and the y-coordinate of the centroid is the average of the y-coordinates of the n nearest
neighbours. The coordinates of the centroids return to the client side function and can be used for the
plot of non-disclosive graphs (e.g. scatter plots, heatmap plots, contour plots, etc).
}
\author{
Avraam D
}
