% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDS1.o.R
\name{glmDS1.o}
\alias{glmDS1.o}
\title{glmDS1}
\usage{
glmDS1.o(formula, family, weights, data)
}
\description{
This is the first serverside function called by ds.glm. It is an aggregation function that sets up the model structure
and creates the starting beta.vector that feeds, via ds.glm, into glmDS2 to enable iterative fitting of the generalized linear model that has been
been specified. For more details please see the extensive header for ds.glm.
}
