% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDS2.o.R
\name{glmDS2.o}
\alias{glmDS2.o}
\title{glmDS2.o}
\usage{
glmDS2.o(formula, family, beta.vect, offset, weights, dataName)
}
\description{
This is the second serverside function called by ds.glm. It is an aggregation function that uses the model structure
and starting beta.vector constructed by glmDS1 to iteratively fit the generalized linear model that has been
been specified. The function glmDS2 also carries out a series of disclosure checks and if the arguments or data fail any of those tests,
model construction is blocked and an appropriate serverside error message is created and returned to ds.glm on the clientside.
For more details please see the extensive header for ds.glm.
}
