% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmerSLMADS2.o.R
\name{glmerSLMADS2.o}
\alias{glmerSLMADS2.o}
\title{glmeSLMADS2.o}
\usage{
glmerSLMADS2.o(formula, family, offset, weights, dataName)
}
\description{
This is the second serverside function called by ds.glmSLMA.o.
}
\details{
It is a function
}
