% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmeSLMADS2.o.R
\name{glmeSLMADS2.o}
\alias{glmeSLMADS2.o}
\title{glmeSLMADS2.o}
\usage{
glmeSLMADS2.o(formula, family, offset, weights, dataName)
}
\description{
This is the second serverside function called by ds.glmSLMA.o.
}
\details{
It is a function
}
