% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramDS.o.R
\name{histogramDS.o}
\alias{histogramDS.o}
\title{Computes a histogram of the input variable without plotting.}
\usage{
histogramDS.o(xvect, num.breaks, method.indicator, k, noise)
}
\arguments{
\item{xvect}{the numeric vector for which the histogram is desired.}

\item{num.breaks}{the number of breaks that the range of the variable is divided.}

\item{min}{a numeric, the lower limit of the distribution.}

\item{max}{a numeric, the upper limit of the distribution.}
}
\value{
a list with an object of class \code{histogram} and the number of invalid cells
}
\description{
This function produces the information required to plot
a histogram. This is done without allowing for bins (cells) with number
of counts less than the pre-specified disclosure control set for the minimum cell
size of a table. If a bin has less counts than this threshold then their counts
and its density are replaced by a 0 value.
}
\author{
Amadou Gaye, Demetris Avraam for DataSHIELD Development Team
}
