% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramDS1.o.R
\name{histogramDS1.o}
\alias{histogramDS1.o}
\title{Computes a histogram of the input variable without plotting.}
\usage{
histogramDS1.o(xvect, min, max, num.breaks)
}
\arguments{
\item{xvect}{the numeric vector for which the histogram is desired.}

\item{min}{a numeric, the lower limit of the distribution.}

\item{max}{a numeric, the upper limit of the distribution.}

\item{num.breaks}{the number of breaks that the range of the variable is divided.}
}
\value{
a list with an object of class \code{histogram} and the number of invalid cells
}
\description{
This function produces the information required to plot
a histogram. This is done without allowing for bins (cells) with number
of counts less than the pre-specified disclosure control set for the minimum cell
size of a table. If a bin has less counts than this threshold then their counts 
and its density are replaced by a 0 value.
}
\author{
Gaye A, Avraam D
}
