% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramDS2.o.R
\name{histogramDS2.o}
\alias{histogramDS2.o}
\title{Computes a histogram of the input variable without plotting.}
\usage{
histogramDS2.o(xvect, min, max, num.breaks, num.neighbours)
}
\arguments{
\item{xvect}{the numeric vector for which the histogram is desired}

\item{min}{a numeric, the lower limit of the distribution}

\item{max}{a numeric, the upper limit of the distribution}

\item{num.breaks}{the number of breaks that the range of the variable is divided}

\item{num.neighbours}{the number of the nearest neghbours for which their centroid is calculated}
}
\value{
a list with an object of class \code{histogram} and the number of invalid cells. The
number of invalid cells for the method "Centroids" is set to null.
}
\description{
This function produces the information required to plot
a histogram.
}
\details{
This function produces the information required to plot
a histogram. This is done by locating the centroids for each k nearest neighbours
of the input data and then producing information for the histogram of the centroids.
}
\author{
Avraam D
}
