% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS1.b.R
\name{lexisDS1.b}
\alias{lexisDS1.b}
\title{lexisDS1}
\usage{
lexisDS1.b(exitCol = NULL)
}
\description{
The first serverside function called by ds.lexis. This is an aggregate function which identfies
the maximum surival time in each study. Adds a random increment to that maximum (to protect against
disclosure) and then passes these maxima back to ds.lexis. ds.lexis then identifies the maximum of the
maxima and this allows a coherent common set of time intervals to be set across all studies when
the collapsed dataframe is expanded by lexisDS2, usually in preparing for a piecewise exponential
regression. For more details see the extensive header for ds.lexis.
}

