% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS1.o.R
\name{lexisDS1.o}
\alias{lexisDS1.o}
\title{lexisDS1}
\usage{
lexisDS1.o(exitCol = NULL)
}
\description{
The first serverside function called by ds.lexis. This is aggregate function.
For more details see the extensive header for ds.lexis.
}
