% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS2.b.R
\name{lexisDS2.b}
\alias{lexisDS2.b}
\title{lexisDS2}
\usage{
lexisDS2.b(datatext = NULL, intervalWidth, maxmaxtime, idCol, entryCol,
  exitCol, statusCol, vartext = NULL)
}
\description{
The second serverside function called by ds.lexis. This is the assign function which actually creates
the expanded dataframe containing surival data for a piecewise exponential regression. lexisDS2 also
carries out a series of disclosure checks and if the arguments or data fail any of those tests,
creation of the exapanded dataframe is blocked and an appropriate serverside error message is stored.
For more details see the extensive header for ds.lexis.
}

