% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS3.o.R
\name{lexisDS3.o}
\alias{lexisDS3.o}
\title{lexisDS3}
\usage{
lexisDS3.o()
}
\description{
The third serverside function called by ds.lexis. This is an assign function that simplifies the
returned output from ds.lexis. Specifically, without lexisDS3 the output consists of a table within
a list, but lexisDS3 converts this directly into a dataframe.
For more details see the extensive header for ds.lexis.
}
