% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSLMADS2.o.R
\name{lmerSLMADS2.o}
\alias{lmerSLMADS2.o}
\title{lmeSLMADS2.o}
\usage{
lmerSLMADS2.o(formula, offset, weights, dataName, REML = TRUE)
}
\description{
This is the second serverside function called by ds.glmSLMA.o.
}
\details{
It is a function
}
