% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeSLMADS2.o.R
\name{lmeSLMADS2.o}
\alias{lmeSLMADS2.o}
\title{lmeSLMADS2.o}
\usage{
lmeSLMADS2.o(formula, offset, weights, dataName)
}
\description{
This is the second serverside function called by ds.glmSLMA.o.
}
\details{
It is a function
}
