% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageDS.o.R
\name{messageDS.o}
\alias{messageDS.o}
\title{messageDS.o}
\usage{
messageDS.o(message.object.name)
}
\arguments{
\item{message.obj}{is a character string, containing the name of the list containing the
message. See the header of the client-side function ds.message.o for more details.
#' @return a list object from each study, containing whatever message has been written by
DataSHIELD into $studysideMessage.}
}
\description{
This function allows for error messages arising from the 
running of a server-side assign function to be returned to the client-side
}
\author{
Burton PR
}
