% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeDS.o.R
\name{rangeDS.o}
\alias{rangeDS.o}
\title{returns the minimum and maximum of a numeric vector}
\usage{
rangeDS.o(xvect)
}
\arguments{
\item{xvect}{a numerical}
}
\value{
a numeric vector which contains the minimum and the maximum
}
\description{
this function is similar to R function \code{range} but instead to return 
the real minimum and maximum, the computed values are multiplied by a very small random number.
}
\details{
This functions finds the minimum and the maximum of the input numeric variable and 
adds to them a random number. It then returns the outcome to the client side.
}
\author{
Gaye A, Avraam D
}
