% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reShapeDS.o.R
\name{reShapeDS.o}
\alias{reShapeDS.o}
\title{reShapeDS.o called by ds.reShape.o}
\usage{
reShapeDS.o(data.name, varying.transmit, v.names.transmit, timevar.name,
  idvar.name, drop.transmit, direction, sep)
}
\arguments{
\item{data.name, }{the name of the data frame.}

\item{varying.transmit, }{names of sets of variables in the wide format that correspond to single
variables in long format (‘time-varying’).}

\item{v.names.transmit, }{the names of variables in the long format that correspond to multiple variables
in the wide format}

\item{timevar.name, }{the name of the variable in long format that differentiates multiple records from the same
group or individual. If more than one record matches, the first will be taken}

\item{idvar.name, }{Names of one or more variables in long format that identify multiple records from
the same group/individual. These variables may also be present in wide format.}

\item{drop.transmit, }{a vector of names of variables to drop before reshaping.}

\item{direction, }{a character string, matched to either "wide" to reshape to wide 
format, or "long" to reshape to long format.}

\item{sep, }{a character vector of length 1, indicating a separating character in the variable
names in the wide format. This is used for guessing v.names and times arguments based on the
names in varying. This is also used to create variable names when reshaping to wide format.}
}
\value{
a dataframe will be created on the serverside and named according to the \code{newobj} argument of the
clientside function ds.reShape.o()
}
\description{
An assign serverside function that reshapes a given data frame between wide and long formats.
}
\details{
This function reshapes a data frame between ‘wide’ format with repeated
measurements in separate columns of the same record and ‘long’ format with the repeated
measurements in separate records and assign it at the servers.
}
\author{
Demetris Avraam, Paul Burton for DataSHIELD Development Team
}
