% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rilmDS.b.R
\name{rilmDS.b}
\alias{rilmDS.b}
\title{Generates matrices for the sum of squares}
\usage{
rilmDS.b(..., yvar)
}
\description{
This function creates matrices having elements the sum of squares that will be used on random intercept linear model method.
}
\details{
This function returns a list with four elements. The first is the length of a variable's vector when we remove from the data frame any rows with NAs.
}
