% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDS.o.R
\name{rmDS.o}
\alias{rmDS.o}
\title{Perform removal}
\usage{
rmDS.o(x.name.transmit)
}
\value{
list of removed objects
}
\description{
Perform removal.
}
\details{
Remove R object from server.
}
\author{
Burton PR
}
