% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDS.o.R
\name{seqDS.o}
\alias{seqDS.o}
\title{seqDS.o called by ds.seq.o}
\usage{
seqDS.o(FROM.value.char, BY.value.char, LENGTH.OUT.value.char,
  ALONG.WITH.name)
}
\arguments{
\item{FROM.value.char, }{a number given as a character denoting the starting value of the sequence. 
The default value is set to 1.}

\item{BY.value.char, }{a number given as a character indicating the increment of the sequence.
The default value is set to 1.}

\item{LENGTH.OUT.value.char, }{a non-negative number given as character denoting the desired 
length of the sequence.}

\item{ALONG.WITH.name, }{is the name of a serverside vector in inverted commas that is used to
determine the length of the created vector.}
}
\value{
a sequence vector will be created on the serverside and named according to the
\code{newobj} argument of the clientside function ds.seq.o()
}
\description{
An assign serverside function that generates sequence vectors.
}
\details{
An assign function that uses the native R function seq() to create
any one of a flexible range of sequence vectors that can then be used to help
manage and analyse data.
}
\author{
Paul Burton, Demetris Avraam for DataSHIELD Development Team
}
