% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2DDS.o.R
\name{table2DDS.o}
\alias{table2DDS.o}
\title{Computes statistical mean of a vectores}
\usage{
table2DDS.o(xvect, yvect)
}
\arguments{
\item{xvect}{a vector}

\item{yvect}{a vector}
}
\value{
output table and message
}
\description{
Some text.
}
\details{
Some text.
}
\author{
Avraam D., Burton PR.
}
