% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapplyDS.assign.o.R
\name{tapplyDS.assign.o}
\alias{tapplyDS.assign.o}
\title{tapplyDS.assign.o called by ds.tapply.assign.o}
\usage{
tapplyDS.assign.o(X.name, INDEX.names, FUN.name)
}
\arguments{
\item{X.name, }{the name of the variable to be summarized. The user must set the name as a
character string in inverted commas}

\item{INDEX.names, }{the name of a single factor or a list of factors to index the variable 
to be summarized. The user must specify this argument in inverted commas.}

\item{FUN.name, }{the name of a valid function to be applied. The present version of this
function allows the user to choose one of the four main summarizing functions that are 'mean',
'sd', 'sum', or 'quantile'.}
}
\value{
an array will be created on the serverside and named according to the
\code{newobj} argument of the clientside function ds.tapply.assign.o()
}
\description{
An assign serverside function that applies a function over a ragged array.
}
\details{
An assign function that uses the native R function tapply() to summarize a variable
over one or more factors
}
\author{
Paul Burton, Demetris Avraam for DataSHIELD Development Team
}
