% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapplyDS.o.R
\name{tapplyDS.o}
\alias{tapplyDS.o}
\title{tapplyDS.o called by ds.tapply.o}
\usage{
tapplyDS.o(X.name, INDEX.names, FUN.name)
}
\arguments{
\item{X.name, }{the name of the variable to be summarized. The user must set the name as a
character string in inverted commas}

\item{INDEX.names, }{the name of a single factor or a list of factors to index the variable 
to be summarized. The user must specify this argument in inverted commas.}

\item{FUN.name, }{the name of a valid function to be applied. The present version of the tapplyDS.o
function allows the user to choose one of the four main summarizing functions that are the 'mean',
the 'sd', the 'sum', and the 'quantile'.}
}
\value{
an array will be returned to the clientside
}
\description{
An aggregate serverside function that applies a function over a ragged array.
}
\details{
An aggregate function that uses the native R function tapply() to summarize a variable
over one or more factors
}
\author{
Paul Burton, Demetris Avraam for DataSHIELD Development Team
}
