% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testObjExistsDS.o.R
\name{testObjExistsDS.o}
\alias{testObjExistsDS.o}
\title{testObjExistsDS}
\usage{
testObjExistsDS.o(test.obj.name = NULL)
}
\description{
This is aggregate function.
}
\details{
For more details see the extensive header for ds.testObjExists
}
