% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varDS.o.R
\name{varDS.o}
\alias{varDS.o}
\title{Computes the variance of vector}
\usage{
varDS.o(xvect)
}
\arguments{
\item{xvect}{a vector}
}
\value{
a numeric, the variance
}
\description{
Calculates the variance.
}
\details{
if the length of input vector is less than the set filter
a missing value is returned.
}
\author{
Amadou Gaye, Demetris Avraam, for DataSHIELD Development Team
}
