% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS1.o.R
\name{lexisDS1.o}
\alias{lexisDS1.o}
\title{lexisDS1}
\usage{
lexisDS1.o(exitCol = NULL)
}
\arguments{
\item{exitCol}{a character string specifying the variable holding the time that each individual is censored or fails}
}
\description{
The first serverside function called by ds.lexis. This is an aggregate function.
For more details see the extensive header for ds.lexis.
}
\author{
Burton PR
}
