% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeValuesDS1.o.R
\name{recodeValuesDS1.o}
\alias{recodeValuesDS1.o}
\title{recodeValuesDS1}
\usage{
recodeValuesDS1.o(var.name.text = NULL, values2replace.text = NULL,
  new.values.text = NULL)
}
\description{
The first serverside function called by ds.recodeValues. This is aggregate function.
}
\details{
For more details see the extensive header for ds.recodeValues.
}
