% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeValuesDS2.o.R
\name{recodeValuesDS2.o}
\alias{recodeValuesDS2.o}
\title{recodeValuesDS2}
\usage{
recodeValuesDS2.o(var.name.text = NULL, values2replace.text = NULL,
  new.values.text = NULL, numeric.output.format.possible,
  force.output.format = "no", v2r.numeric = NULL)
}
\description{
The second serverside function called by ds.recodeValues. This is assign function.
}
\details{
For more details see the extensive header for ds.recodeValues.
}
