% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rilmDS.b.R
\name{rilmDS.b}
\alias{rilmDS.b}
\title{Generates matrices for the sum of squares}
\usage{
rilmDS.b(..., yvar)
}
\arguments{
\item{..., }{any explanatory variables}

\item{yvar, }{the outcome variable}
}
\value{
(i) the number of complete cases, (ii) the sum of squares of 
explanatory variables, (iii) the sum of products of each explanatory
variable with the outcome variable, and (iv) the sum of squares of the
outcome variable
}
\description{
This function creates matrices having elements the sum of
squares that will be used on random intercept linear model method.
}
\details{
This function returns a list with four elements. The first is the
length of a variable's vector when we remove from the data frame any rows
with NAs.
}
\author{
Avraam D; Jones EM; van den Heuvel ER
}
